+++ cryptex/play.php
@@ -89,6 +89,8 @@ function game_cryptex_continue( $cm, $game, $attempt, $cryptexrec, $endofgame, $
 
     // The game->param4 is minimum words.
     // The game->param2 is maximum words.
+    $badwords = array( 'NO', 'ON', 'ASS', 'SSA', 'SEX', 'XES');
+    $cryptex->setbadwords( $badwords);
     if ($cryptex->computedata( $crossm, $crossd, $letters, $game->param4, $game->param2, $game->param3)) {
         $newcrossd = array();
         foreach ($crossd as $rec) {